Crate panic_hook

A simple panic hook that allows skipping printing stack trace conditionally.

This crate is only for use during testing - in production, any panic in TiKV is fatal, and it is compiled with panic=abort, so use of this crate in production is an error.

Constants

MUTED

Statics

DEFAULT_HOOK

INIT

Functions

initialize
Replace the default hook if we haven't.

mute
Skip printing the stack trace if panic.

recover_safe
Recover from closure which may panic.

track_hook
Print the stacktrace according to the static MUTED.

unmute
Print the stack trace if panic.